Add documentation and adjust `kwargs`
Added documentation using `Documenter.jl`

Adjusted `kwargs` to allow keywords to be specified in the original way
as well as how `args` can be specified: i.e. in a sub-configuration
dictionary. In such a case, to give a keyword for a constructor at level
`[1.x]` you could write `[1.x.":keyword"]`.